Some minor code refactoring. Adding insights functionality.
