+++ src/cli.py
@@ -5,7 +5,7 @@ import argparse
 import sys
 
 ### Argparse text
-usage = "track.py [-a COMPANY_NAME] [-u COMPANY_NAME] [-d COMPANY_NAME] [-l]"
+usage = "track.py [-a COMPANY_NAME] [-u COMPANY_NAME] [-d COMPANY_NAME] [-l] [-i]"
 description = "Struggle Tracker - A program that helps you track your job applications"
 epilog = r"""
 LIST OPTIONS
@@ -15,6 +15,14 @@ LIST OPTIONS
   title          sort by job title
   status         sort by status
   notes          sort by notes
+
+INSIGHT OPTIONS
+  all            print all job status insights
+  pending        print percentage of pending job applications
+  in_progress    print percentage of in progress job applications
+  offers         print percentage of job applications with an offer
+  hired          print percentage of jobs you have been hired at
+  rejected       print percentage of rejected job applications
 
 EXAMPLES
 
@@ -47,6 +55,7 @@ def parse_args():
     scraper.add_argument("-u","--update",nargs=1,metavar="",help="update an existing company in the spreadsheet")
     scraper.add_argument("-d","--delete",nargs=1,metavar="",help="delete an existing company in the spreadsheet")
     scraper.add_argument("-l","--list",const="date",default="date",nargs="?",metavar="",help="list all saved job applications")
+    scraper.add_argument("-i","--insights",const="all",default="all",nargs="?",metavar="",help="display job application insights")
 
     if len(sys.argv[1:]) == 0:
         parser.print_help()
