+++ src/functions/insight_functions.py
@@ -1,0 +1,12 @@
+#===============================================================================
+#                                   Insight Functions
+#===============================================================================
+from colorama import Fore, init, Style
+from .. import global_vars
+import csv
+
+init(autoreset=True)
+
+job_categories = global_vars.job_categories
+status_options = global_vars.status_options
+
