+++ src/functions/update_functions.py
@@ -1,9 +1,11 @@
 #===============================================================================
 #                               Update Functions
 #===============================================================================
-from colorama import Fore, Style
+from colorama import Fore, init, Style
 from .. import global_vars
 from . import csv_functions
+
+init(autoreset=True)
 
 job_categories = global_vars.job_categories
 status_options = global_vars.status_options
