+++ src/programs/insight.py
@@ -1,0 +1,6 @@
+#===============================================================================
+#                       Return Job Application Insights
+#===============================================================================
+
+def print_insights(args,insight_functions):
+    pass
